refactor(evm): move the execution traits to `execution.rs`
